% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.crf.R
\name{predict.crf}
\alias{predict.crf}
\title{Predictions from a crf given newdata}
\usage{
\method{predict}{crf}(object, newdata, sderr = FALSE, ...)
}
\arguments{
\item{object}{a fitted \code{crf} clustered random forest object fitted by \code{crf}.}

\item{newdata}{dataset on which predictions are to be performed.}

\item{sderr}{whether 'bootstrap of little bags' standard errors should be additionally outputted. Default is \code{FALSE}.}

\item{...}{additional arguments}
}
\value{
Fitted values, potentially alongside standard errors (see \code{sderr}).
}
\description{
Predictions from a fitted \code{crf} clustered random forest on newdata \code{newdata}.
}
