% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.crf}
\alias{summary.crf}
\title{Summary for a crf fitted object}
\usage{
\method{summary}{crf}(object, ...)
}
\arguments{
\item{object}{a fitted \code{crf} clustered random forest object fitted by \code{crf}.}

\item{...}{additional arguments}
}
\value{
Prints summary output for \code{crf} object
}
\description{
Summary of a fitted \code{crf} clustered random forest object fitted by \code{crf}.
}
